course page convert into SSR
